pyproject.toml
[tool.poetry]
name = "botnet-detector"
version = "0.1.0"
description = "AWS CTI"
authors = ["Craftech"]
license = "MIT"
readme = "README.md"
packages = [{include = "botnet_detector"}]

[tool.poetry.dependencies]
python = "^3.11"
stix-shifter = "^5.3.0"
stix-shifter-utils = "^5.3.0"
stix-shifter-modules-aws-athena = "^5.3.0"
stix-shifter-modules-aws-cloud-watch-logs = "^5.3.0"
stix2 = "^3.0.1"
python-dotenv = "^1.0.0"
jsonlines = "^3.1.0"
kestrel-lang = "^1.6.0"
stix2-validator = "^3.1.3"
kestrel-jupyter = "^1.0.8"
stix-shifter-modules-elastic-ecs = "^5.3.0"
tqdm = "^4.65.0"
numpy = "^1.24.3"
matplotlib = "^3.7.1"
geoip2 = "^4.7.0"
folium = "^0.14.0"
gower = "^0.1.2"
sklearn = "^0.0.post5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
